combine token pair into one topic, add timestamps
Since we can only have three indexes per event, I had to leave out the
index on the taker address before in order to fit both token addresses.
We can overcome this by combining the token addresses into one index.

(This changes only the new events introduced in the last update, so the
frontends should not be affected.)